+++ lib/kitchen/directions/bake_free_response/main.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module BakeFreeResponse
+      def self.v1(chapter:, metadata_source:, append_to: nil)
+        V1.new.bake(chapter: chapter, metadata_source: metadata_source, append_to: append_to)
+      end
+    end
+  end
+end